+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "server",
     "client",
+    "common",
 ]
 
 resolver = "2"
