+++ common/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "common"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+path = "lib.rs"
+[dependencies]
+bincode = "1.3.3"
+serde = {version = "1.0.210",features = ["derive"]}
