+++ server/Cargo.toml
@@ -12,6 +12,7 @@ log = "0.4.22"
 serde = {version = "1.0.210",features = ["derive"]}
 tokio = { version = "1.40.0", features = ["full"] }
 tokio-tungstenite = "0.24.0"
+common = { path = "../common" }
 
 
 [dev-dependencies]
